Recieve FileName Plus Extension From User
At this point, the system promts the user for:
1). Enter the number of living and office rooms
2). enter the file name plus a .txt extension

after entering data the system generates room names and automatically
allocates the users to appreopriate rooms randomly.

it also prints the alocations to a text file called output.

Add Listings Of Allocated And Unallocated People

Implemented the printing of un allocated  and allocated fellows and staff
it writes out to a file(output.txt) under the data folder but it over-writes
the same file.

Future work is to implement a method that checks for how many rooms are unallocated
and append to that data